fix: clear stale turn errors on retry; add Idempotency-Key support
- A retryable error_message/loop_error left recordedError set, so a
  turn that retried and succeeded still settled as a 500. New assistant
  output now clears the recorded error (errors after the final text
  still fail the request).
- Honor Idempotency-Key / X-Idempotency-Key: retries carrying the same
  key share the original request's outcome (in-flight duplicates join
  the same turn; replays stream the cached text as one chunk) instead
  of appending the user message again. Failed outcomes are evicted so
  deliberate retries re-run. Bounded FIFO cache.
- Split turn execution into app-server-openai-turn.ts (1000-line file
  limit); no behavior change from the split.
